backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
pandas==2.1.3
openpyxl==3.11.0
qrcode==7.4.2
Pillow==10.1.0
pydantic==2.5.0
email-validator==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.8.1
bcrypt==4.1.2
httpx==0.25.2
requests==2.31.0
